Add convenience methods for programmatically switching between in-memory and local lookup tables. Move implementations to internal.
